Feature 50849, reduce_by_writable_permissions shortcut for new models
For new models, per-field permissions are not considered, only the
default permission (e.g. add_work_package).
This is may not be a final solution, but at this moment I don't know how
to distinguish between per-field permissions for edit and create cases.

This fixes more tests, but copy WP scenario required also explicit
`status_id` in the merge block in CopyService#copied_attributes.
This was previously done for `parent_id` as well, so it's not exactly
a precedent.

+++ app/contracts/base_contract.rb
@@ -242,6 +242,13 @@ class BaseContract < Disposable::Twin
   def reduce_by_writable_permissions(attributes) # rubocop:disable Metrics/PerceivedComplexity, Metrics/AbcSize
     attribute_permissions = collect_ancestor_attributes(:attribute_permissions)
 
+    # For new models we don't go to per-field checks if the user has any of the default permissions.
+    permissions = attribute_permissions[:default_permission]
+    if model.new_record? && permissions
+      return attributes if model.project.present? && permissions.any? { |perm| user.allowed_in_project?(perm, model.project) }
+      return attributes if model.project.blank? && permissions.any? { |perm| user.allowed_in_any_project?(perm) }
+    end
+
     attributes.reject do |attribute|
       canonical_attribute = attribute.delete_suffix('_id')
